+++ Mouse Tracking/viola-jons/AdaBoost.h
@@ -10,17 +10,17 @@ class Learner;
 class AdaBoost
 {
 public:
-    vector<vector<double>> X;
+    vector<vector<int>> X;
     vector<int> y;
     vector<double> weights;
     vector<Learner *> learners;
     vector<double> alphas;
-    AdaBoost(vector<vector<double>> X, vector<int> y);
+    AdaBoost(vector<vector<int>> X, vector<int> y);
     AdaBoost();
     ~AdaBoost();
 
     void train(int T);
-    int predict(const std::vector<double> &X);
+    int predict(const std::vector<int> &X);
     void save(const string file);
     void load(const string file);
     void saveAsText(const string &file);
